Use material components and alter one by altering one of the parameters
In the AppTheme, add textAppearanceHeadline6 and make it reference
another custom style. Inside there, make sure that the parent is Base.XX
where XX is whatever we are trying to modify. And then simply add
an attribute, which will override the ones specified higher up in the
hierarchy.

To ensure that the styles are applied correctly and nothing is
overridden, we must use android:textAppearance on the TextViews instead
of style directly. We then reference the correct style by providing a
question mark and then the style we want to apply, like this:
(android:textAppearance="?textAppearanceHeadline6").