Override the theme on a specific view with a theme overlay.
Alter Toolbar theme specifically, by applying the Dark theme just on it.
This is also applied on all the child views.

This does not change the color of the ToolBar directly so we do that
manually. We also apply a tint on the child image to look better on top
of the new primary color.